Split mpdevil.py into separate modules
