+++ src/mpdevil/gui/main_window/__init__.py
@@ -1,0 +1,1 @@
+from mpdevil.gui.main_window.main_window import MainWindow
