+++ src/mpdevil/gui/main_window/browser/__init__.py
@@ -1,0 +1,1 @@
+from mpdevil.gui.main_window.browser.browser import Browser
